+++ include/os_scheduler.h
@@ -71,7 +71,7 @@ void os_print_utilization(void);
  * @param uart_in Zeiger auf Ausgabeparameter fuer UART-Auslastung eingehend
  * @param uart_out Zeiger auf Ausgabeparameter fuer UART-Auslastung ausgehend
  */
-void os_get_utilizations(uint8_t* cpu, uint16_t* uart_in, uint16_t* uart_out);
+void os_get_utilizations(uint8_t* cpu, uint8_t* uart_in, uint8_t* uart_out);
 
 /**
  * Handler fuer OS-Display
